liaopeng:1月23号，更改退货退捐（因前期实现搞错了），更改未确认收货不进入结算，更改退货订单结算金额减去退货的进货价，更改捐增数额保留小数点两位，更改总后台消费类型选择的界面问题，更改退货列表中的价格及数量显示问题。

+++ Application/Admin/Controller/GoodsController.class.php
@@ -523,7 +523,7 @@ class GoodsController extends MallController{
         $data['is_cash'] = isset($_POST['is_cash']) ? 1: 0;//是否允许现金
         $data['is_yqt'] =  isset($_POST['is_yqt']) ? 1: 0;//是否允许一券通
         $data['is_gwq'] = isset($_POST['is_gwq']) ? 1: 0;//是否允许购物券
-        $data['love_amount'] = $_POST['love_amount'] ? intval($_POST['love_amount']): 0;// 捐赠的数额
+        $data['love_amount'] = $_POST['love_amount'] ? floatval($_POST['love_amount']): 0;// 捐赠的数额
 
         //处理促销时间
         $promote_start_date = empty(I('post.promote_start_date')) ? 0 : I('post.promote_start_date');
